Improved headings in README.md file
Normally heading level 1 is reserved for the document's title, every
heading following is lower than the title heading (H1). I went ahead
and set the README's title to H1 and every following section to it's
respecting heading level. Also the title of the document was using the
"========" bar to set it as H1 while the rest of the headings were
using the "##" prefix to set the heading to H2,H3. I edited the
title's heading so it will be uniform with the rest of the document.